Add BinaryStringTestSupport
