+++ Consent String SDK SwiftTests/Consent_String_SDK_SwiftTests.swift
@@ -9,7 +9,7 @@
 import XCTest
 @testable import Consent_String_SDK_Swift
 
-class Consent_String_SDK_SwiftTests: XCTestCase {
+class Consent_String_SDK_SwiftTests: XCTestCase, BinaryStringTestSupport {
     
     let base64 = ["BOMexSfOMexSfAAABAENAA////ABSABgACAAIA",
                   "BOMexSfOMexSfAAABAENAA////ABSABgACBAAA",
@@ -330,47 +330,5 @@ class Consent_String_SDK_SwiftTests: XCTestCase {
             XCTAssert(consentString.consentScreen == values[index], "Actual screen: \(consentString.consentScreen)")
         }
     }
-    
-    func binaryStringRepresenting(data:Data) -> String {
-        return  data.reduce("") { (acc, byte) -> String in
-            let stringRep = String(byte, radix: 2)
-            let pad = 8 - stringRep.count
-            let padString = "".padding(toLength: pad, withPad: "0", startingAt: 0)
-            return acc + padString + stringRep
-        }
-    }
-    
-    func binary(string:String, isEqualToBinaryString string2:String) -> Bool {
-        if abs(string.count - string2.count) > 7 {
-            return false
-        }
-        var index = 0
-        var max = string.count
-        if string.count > string2.count {
-            max = string2.count
-        }
-        while index < max {
-            if string[string.index(string.startIndex, offsetBy: index)] != string2[string2.index(string2.startIndex, offsetBy: index)] {
-                return false
-            }
-            index += 1
-        }
-        if string.count > string2.count {
-            while index < string.count {
-                if string[string.index(string.startIndex, offsetBy: index)] != "0" {
-                    return false
-                }
-                index += 1
-            }
-        } else {
-            while index < string2.count {
-                if string2[string2.index(string2.startIndex, offsetBy: index)] != "0" {
-                    return false
-                }
-                index += 1
-            }
-        }
-        return true
-    }
-    
+
 }
